git-svn-id: https://scm.gforge.inria.fr/authscm/jmcoutey/svn/pm2/trunk@17968 ee107e0b-c2fa-0310-842a-ec2723fb34f4

+++ nmad/schedulers/sched_opt/src/nm_so_strat_aggreg.c
@@ -365,8 +365,8 @@ static int exit_strategy(void)
  *  @return The NM status.
  */
 static int rdv_accept(struct nm_gate *p_gate,
-		      unsigned long *drv_id,
-		      unsigned long *trk_id)
+		      uint8_t *drv_id,
+		      uint8_t *trk_id)
 {
   struct nm_so_gate *p_so_gate = p_gate->sch_private;
