+++ nmad/schedulers/sched_opt/src/nm_so_strat_qos.c
@@ -66,7 +66,7 @@ static int pack_ctrl(struct nm_gate *p_gate,
 
   err = policies[policy]->pack_ctrl(p_so_sq_gate->policy_priv[policy],
 				    p_ctrl);
-  
+
   return err;
 }
 
@@ -104,7 +104,7 @@ static int get_next_busy_policy(struct nm_so_strat_qos_gate *p_so_sq_gate)
       policy = (policy + 1) % NM_SO_NB_POLICIES;
       round = policies_weights[policy];
     }
-  
+
   for(i = 0; i < NM_SO_NB_POLICIES; i++)
     {
       round--;
@@ -114,12 +114,12 @@ static int get_next_busy_policy(struct nm_so_strat_qos_gate *p_so_sq_gate)
 	  p_so_sq_gate->current_policy = policy;
 	  return policy;
 	}
-      
+
       policy = (policy + 1) % NM_SO_NB_POLICIES;
       round = policies_weights[policy];
-      
+
     }
-  
+
   return NM_SO_POLICY_UNDEFINED;
 }
 
@@ -133,7 +133,7 @@ static int try_and_commit(struct nm_gate *p_gate)
     = (struct nm_so_strat_qos_gate *)p_so_gate->strat_priv;
   int err = NM_ESUCCESS;
   uint8_t policy;
-   
+
   if(p_so_gate->active_send[NM_SO_DEFAULT_NET][TRK_SMALL] ==
      NM_SO_MAX_ACTIVE_SEND_PER_TRACK)
     /* We're done */
@@ -142,11 +142,11 @@ static int try_and_commit(struct nm_gate *p_gate)
   if((policy = get_next_busy_policy(p_so_sq_gate)) == NM_SO_POLICY_UNDEFINED)
     /* No data available */
     goto out;
-  
-  err = 
+
+  err =
     policies[policy]->try_and_commit(p_gate,
 				     p_so_sq_gate->policy_priv[policy]);
- out:  
+ out:
   return err;
 }
 
@@ -159,22 +159,22 @@ static int init(void)
   policies[NM_SO_POLICY_LATENCY] = &nm_so_policy_latency;
   policies_weights[NM_SO_POLICY_LATENCY] = 1;
   policies[NM_SO_POLICY_RATE] = &nm_so_policy_rate;
-  policies_weights[NM_SO_POLICY_RATE] = 1;  
+  policies_weights[NM_SO_POLICY_RATE] = 1;
   policies[NM_SO_POLICY_PRIORITY] = &nm_so_policy_priority;
   policies_weights[NM_SO_POLICY_PRIORITY] = 1;
   policies[NM_SO_POLICY_PRIORITY_LATENCY] = &nm_so_policy_priority_latency;
   policies_weights[NM_SO_POLICY_PRIORITY_LATENCY] = 1;
   policies[NM_SO_POLICY_PRIORITY_RATE] = &nm_so_policy_priority_rate;
   policies_weights[NM_SO_POLICY_PRIORITY_RATE] = 1;
-  
+
   return NM_ESUCCESS;
 }
 
 /* Warning: drv_id and trk_id are IN/OUT parameters. They initially
    hold values "suggested" by the caller. */
 static int rdv_accept(struct nm_gate *p_gate,
-		      unsigned long *drv_id,
-		      unsigned long *trk_id)
+		      uint8_t *drv_id,
+		      uint8_t *trk_id)
 {
   struct nm_so_gate *p_so_gate = p_gate->sch_private;
 
@@ -198,17 +198,17 @@ static int ack_callback(struct nm_so_pkt_wrap *p_so_pw,
   struct nm_so_strat_qos_ack *p_so_ack;
   uint8_t tag = tag_id - 128;
   uint8_t i;
-  
+
   if(!finished)
     {
       uint8_t priority = nm_so_get_priority(tag);
-      
+
       p_so_ack = tbx_malloc(p_so_sq_gate->nm_so_ack_mem);
-     
+
       p_so_ack->tag = tag;
       p_so_ack->seq = seq;
       p_so_ack->track_id = track_id;
-      
+
       list_add_tail(&p_so_ack->link, &p_so_sq_gate->pending_acks[priority]);
     }
   else
@@ -220,27 +220,27 @@ static int ack_callback(struct nm_so_pkt_wrap *p_so_pw,
 	      tag = p_so_ack->tag;
 	      seq = p_so_ack->seq;
 	      track_id = p_so_ack->track_id;
-	      
+
 	      NM_SO_TRACE("ACK completed for tag = %d, seq = %u\n", tag, seq);
-	     
+
 	      p_so_gate->pending_unpacks--;
-	      
+
 	      list_for_each_entry(p_so_large_pw, &p_so_gate->pending_large_send[tag], link)
 		{
-		  if(p_so_large_pw->pw.seq == seq) 
+		  if(p_so_large_pw->pw.seq == seq)
 		    {
 		      list_del(&p_so_large_pw->link);
-		      
+
 		      /* Send the data */
 		      _nm_so_post_send(p_gate, p_so_large_pw,
 				       track_id % NM_SO_MAX_TRACKS,
 				       track_id / NM_SO_MAX_TRACKS);
-		     
+
 		      goto next;
-		    }	    
+		    }
 		}
 	      TBX_FAILURE("PANIC!\n");
-	      
+
 	    next:
 	      list_del(&p_so_ack->link);
 	      tbx_free(p_so_sq_gate->nm_so_ack_mem, p_so_ack);
@@ -255,7 +255,7 @@ static int init_gate(struct nm_gate *p_gate)
   uint8_t i;
   struct nm_so_strat_qos_gate *priv
     = TBX_MALLOC(sizeof(struct nm_so_strat_qos_gate));
-  
+
   for(i = 0; i < NM_SO_NB_POLICIES; i++)
     policies[i]->init_gate(&priv->policy_priv[i]);
 
@@ -264,11 +264,11 @@ static int init_gate(struct nm_gate *p_gate)
 
   for(i = 0; i < NM_SO_NB_PRIORITIES; i++)
     INIT_LIST_HEAD(&priv->pending_acks[i]);
-  
+
   tbx_malloc_init(&priv->nm_so_ack_mem,
 		  sizeof(struct nm_so_strat_qos_ack),
 		  16, "nmad/.../sched_opt/nm_so_strat_qos/ack");
-  
+
   ((struct nm_so_gate *)p_gate->sch_private)->strat_priv = priv;
 
   return NM_ESUCCESS;
@@ -279,10 +279,10 @@ static int exit_gate(struct nm_gate *p_gate)
   uint8_t i;
   //  char host_name[30];
   struct nm_so_strat_qos_gate *priv = ((struct nm_so_gate *)p_gate->sch_private)->strat_priv;
-  
+
   for(i = 0; i < NM_SO_NB_POLICIES; i++)
     policies[i]->exit_gate(&priv->policy_priv[i]);
-  
+
   tbx_malloc_clean(priv->nm_so_ack_mem);
 
   TBX_FREE(priv);
